+++ src/roswire/proxy/launch/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-"""This module provides :class:`LaunchFileReader`, a class for parsing the
-contents of XML launch files.
-"""
-from .reader import LaunchFileReader