+++ src/roswire/proxy/roslaunch/config/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+"""
+This module provides data structures for representing roslaunch configurations.
+"""
+from .launch import LaunchConfig
+from .node import NodeConfig
+from .parameter import Parameter